Source code html
Step 1 (HTML Code):

To get started, we will first need to create a basic HTML file. In this file, we will include the main structure for our YouTube clone.

After creating the files just paste the following codes into your file. Make sure to save your HTML document with a .html extension, so that it can be properly viewed in a web browser.

The code begins with the html element, which specifies that the language is English. The head element contains metadata, such as character encoding, viewport, and links to external stylesheets and fonts. The body element contains the main content of the webpage.

Inside the body element, there is a nav element that contains the navigation bar. It includes a menu button, a YouTube logo, a search field with a search button and a voice button, and other navigation options, such as create, YouTube apps, notifications, and user avatar.

Next, there is a div element with the class name chips, which contains a list of categories or tags. Each category is displayed as a div element with the class name chips__item. The selected category has an additional class name of chips__item--selected.

Finally, there is a div element with the class name sidebar that contains a list of links to different sections of the website. Each link is displayed as an a element with the class name sidebar__section__link. The selected link has an additional class name of sidebar__section__link--selected.